Remove splunkconnector entirely.
We're no longer going to log into splunk and create a new
app token. From now on we create the token separately on splunk,
then pass it in through the Jenkins global configuration page. This not
only makes the plugin simpler, but is also how httpinput was intended
to be used.