Remove 'severity' and change Junit error handling logic.
Removed the severity field since it was left over from when splunkins
used javalogging library and because severity made no sense in the
context of Jenkins builds.

Instead of creating a fake junit file. We now have an "error" keypair
which has two sub-keypairs "type" and "message". This makes the
eventdata much cleaner and easier to use. It means Splunkins behaves
more like a user would expect.